[Gradle] Fix MppPublicationCompatibilityIT error in test data regeneration
For some reason when file is missing KotlinTestUtils.assertEqualsToFile
KtUsefulTestCase class can't be initialized by JVM, due to some weird
error with Java Modules.

However, we do have simillar assertion util already.
So I used it and tweaked it a bit so it is compatible with
previous test report format.