[JS] Fix IrCall shape mismatch in ES6ConstructorBoxParameterOptimizationLowering
This lowering used to patch `IrCalls` to ES6 constructor replacement
functions by removing the `box` argument while the shape of the callee
would still expect that argument, thus introducing a shape mismatch.

We fix this by creating new functions and the corresponding IrCalls
without the box parameter. This way the shapes of both always match.

The requirement that shapes must match is important for migrating
to the new function parameter API in the scope of KT-70054.

KT-70803