[IR.Interpreter] Don't store annotations for fake overrides
Annotations for fake overrides have offsets of an original annotation.
Because of that there could be a conflict between the two constants.
One of them can rewrite the entry inside `EvaluatedConstTracker`.
The solution is not to write anything for fake overrides. They are
not going to be serialized anyway.

#KT-72356 Fixed

(cherry picked from commit c312b8697ea4c435f2cf3e98266c34f8a01b7e54)