[K/N] Allow GCStateHolder::schedule in any thread state
`GCStateHolder::schedule` is called from the runnable thread state.
It takes a mutex, which sometimes leads to calling
`std::__1::mutex::lock()`.

This rightfully triggers the thread state checker because this function
can take a long time, which is prohibited in the runnable state.

But in this particular case, the function should be fast, because
`mutex_` is never taken for long there.

To make the thread state checker happy, this commit adds
a `CallsCheckerIgnoreGuard` to `GCStateHolder::schedule`.