Kapt: write stubs before enterTrees in stub converter tests
Apparently, calling `enterTrees` causes javac to insert additional
elements into the tree, such as, for example:
- default constructors
- annotation argument names
- "public static" modifiers for enums in interfaces

This behavior can be disabled inside javac with the flag
`TypeEnter.completionEnabled`, but it seems to be inaccessible from the
outside.

Since kapt in production doesn't call `enterTrees`, it's better if the
stubs test data also wasn't affected by it.