Kapt: do not call enterTrees in stub converter tests
`enterTrees` is not called in kapt in production, so it's better if the
stubs test data also isn't affected by it. In practice, this only leads
to some javac errors being removed on otherwise incorrect code (where
kapt reports errors itself). We may introduce javac run on generated
stubs to catch javac-reported errors separately in the future (related
to KT-71154).

Also, remove the directive `NO_VALIDATION` which is unused now.