[JVM] Fix criterion for choosing between field/getter reflective access
We have to choose between generation of field or getter reflective
access when compiling access to a property with restricted visibility
in a code fragment. Here we simplify the criterion to "always use
getter if it actually exists in bytecode." Also, we take into account
fake overrides and resolve them as it is essential for checking if the
property has explicit accessor.

^KT-72126: Fixed
^KT-72125: Fixed