[K/N] Fix lambdas and function references inside `staticCFunction {}`
Lambdas are lowered to function references early. Function references
are handled by the dedicated function reference lowering.

There is a specific quirk in this lowering needed for `staticCFunction`:
the argument of this function shouldn't be lowered, because it shouldn't
be represented as a class, and `staticCFunction(...)` will be handled as
a direct function pointer at later stages.

Excluding `staticCFunction` argument is achieved by checking the
enclosing IR elements of the `IrFunctionReference` to be lowered: the
check skips the innermost `IrBlock`s and then checks whether the next
element is a proper `IrCall` to `staticCFunction`. In such a case, the
function reference is not lowered.

To check the enclosing IR elements, the lowering puts them into a stack
when traversing the IR tree. The problem is that this stack doesn't
contain declarations. So the check can see a chain of `IrBlock`s,
ignoring a function between them.

Before KT-69549 this wasn't a problem, because local functions were
already lowered before the function references lowering.
But KT-69549 changed that, leading to the bug: function references
are not lowered if they on the top level of a `staticCFunction` lambda:
```kotlin
staticCFunction { p ->
    val lambda = { p } // not lowered.
}
```
The check for `{ p }` just sees the sequence of `IrBlock`s and then the
`IrCall` to `staticCFunction` in the stack, so the lowering skips the
lambda.

This commit fixes that by putting all declarations to the stack as well.

^KT-72552 Fixed

(cherry picked from commit 6918e1ff8b1072ab30631d33e76c590bd9335d84)