[K/N][caches] Fixed multi-process cache build problem
In 5816c0ca2a63d4c9cf05cd13d79b43db70cdc0a5, a synchronization mechanism was introduced to forbid
the same cache to be built simultaneously by a number of processes to reduce the memory usage.
There was a timeout (2 minutes) for the wait for the cache to be built by other process
(which acquired the lock file); unfortunately, this is not enough in some cases (an overloaded machine,
or just a huge library), and the wait might fail.

To overcome this, a new mechanism is introduced in this commit: if a process failed to wait
for the cache to be built successfully by other process, then just build the cache ourselves
(potentially simultaneosly, but, hopefully, it's a rare situation).

 #KT-72281 Fixed