[Analysis API] Load non-JVM libraries with the appropriate platform
For some (perhaps historical) reasons, 'CommonEnvironmentConfigurator'
configures a common stdlib as a `JvmClasspathRoot`. We should not assign
a JVM platform to it, as then the session will have an irrelevant FIR
symbol provider (createStubBasedFirSymbolProviderForClassFiles).

PackagePartProviderTestImpl is adapted as now stdlib in certain tests
(such as FirStandaloneNormalAnalysisSourceModuleCodeFragmentCollectDiagn
osticsTestGenerated.testEvalJvmJs) started loading as a common library,
and type check in 'LLBinaryOriginLibrarySymbolProviderFactory.
createCommonLibrarySymbolProvider' failed. It's unclear, though,
if the 'JvmPackagePartProvider' is even appropriate for non-JVM
environments, but it doesn't seem to be a good replacement on the table.