[Java] Fix DNN computation in enhancement
... in a case when the direct enhancement (from an annotation) is
DNN for warning and some override is enhanced to NOT_NULL for error.
In this case we incorrectly returned DNN with error severity.

After this fix, we enhance to NOT_NULL for error and basically lose
the enhancement to DNN for warning.
The proper solution would be to keep both enhancements (#KT-72620).

#KT-67560 Fixed