minor: Scripting: allow nulls for implicit receivers on eval
while writing a test for script implicit receivers types handling
it was discovered that it is possible to set nullable implicit
receiver type in the compilation configuration, but it is not
allowed to pass null for such a receiver in the evaluation
configuration. This commit fixes it.
One of the tests in the next commit cover the case.