[PSI] introduce `KtFileLikeCodeFragment`
The main target is to serve for Kotlin code injection. Injections can
use imports, package names, top-level expressions, and top-level
declarations. So basically, it's a KtFile but with top-level expressions support.

KtScript does not serve this purpose for multiple reasons:
- The current script considers each script as a module in the project
model, which is too heavyweight for code injection.
- On any host file change (e.g., on a change in a Markdown file that has
 some Kotlin code injected inside), the injected file will be recreated.
  Creating a new script in this case is too heavyweight an operation.
- Scripts operate on physical files, so they require that the script
file is present in the physical file system.