K2 scripting: fix handling of user-provided types
Before this change, all types for script implicit receivers and
provided properties were ad-hoc converted to FirUserTypeRefs and
resolved along the regular resolution. This caused problems with
unresolved types and in some other cases.
The new approach, similarly to K1, tries to find the referenced class
in the dependencies first, and only fall back to creating the
FirUserTypeRef if not found.

#KT-71399 fixed