[K/N][tests] Set timeout for `gsutil` and try repeating if it hits
This command (used for copying the iOS device test execution results
from the cloud) sometimes just hangs on certain CI agents.

This commit sets the timeout for it, and tries to repeat the execution
a couple of times to mitigate this. In a simple experiment, this
little trick proved to work.

^KT-72581 Fixed

(cherry picked from commit 0b9b7dd6928e7800f50de2f14a029dd5fff0194a)