KTIJ-28763 [kotlin] Resolve typealiased constructor references to typealias instead of the underlying constructor
This makes `resolveToSymbols` behavior more aligned with itself,
since it always prefers the type aliases to the underlying symbols

This is most probably not a final design of this issue, since
always resolving to typealias does not seem flexible or
"correct" enough for all possible use-cases.

This change also aligns K1 and K2 implementations of resolve more,
especially in the IDE.
But it's important to note that in K1,
there was a `TypeAliasedConstructorDescriptor`, which had expressed
much more information compared to K2's `KaTypeAliasSymbol`.

There are tickets like KT-60366 and KT-60363 to change this behavior
of references, so that it's possible to resolve such references to
both the typealias and the underlying constructor.

Those tickets are important, but currently have a low priority.
Meanwhile, we need to bring the behavior of references in the IDE
in K2 Mode to at least being on par with K1 Mode.

And always resolving to the underlying constructor seems like a worse
alternative than resolving to the typealias.

^KTIJ-28763 Fixed