[FIR] Replace FirContextReceiver with FirValueParameter
In this commit, FirContextReceiver is removed.
Context receivers are instead represented as FirValueParameters.
The 'name' property of the FirValueParameter replaces
FirContextReceiver.customLabelName and labelNameFromTypeRef.
Because of this, some test data changes were context receivers are
rendered.
Deserialized context receivers always have UNDERSCORE_FOR_UNUSED_VAR as
name, but this has no effect on callsites anyway.
Apart from that, no behavior is changed.

#KT-72222