[K/N][tests] Fix ProgramNameTest with compileOnly=true
13e924e introduced the test for `Platform.programName` from stdlib.
The test checks the API behavior by inspecting a program output. This
doesn't work when running the test in the compile-only mode
(with `-Pkotlin.internal.native.test.compileOnly=true`), because the
executor is no-op then, and it reports empty stdout and stderr.

This commit fixes the problem by making the test not check the output
in the compile-only mode.