[PL] Avoid PL messages for SubclassOptInRequired annotation sites
Don't produce compiler messages when the partial linkage engine
inside the compiler detects that the constructor of the
annotation class SubclassOptInRequired has changed the signature,
and this constructor is used at an annotation site.

However, the messages should still be produced if SubclassOptInRequired
constructor is used as a regular declaration.

^KT-72965 Fixed