Resolution: Fix UNRESOLVED_REFERENCE for callable reference inside elvis
The regression is brought via 20899ecaeba36cb327b543ce10da6c4fdc2556ba
where we assumed that it's fine to use ContextDependent.

The problem was that in some situations (see KT-72996 and KT-73011) the
expected type might be used to resolve callable reference overloads,
as in the original regression where we failed to choose
the correct one.

The idea of the fix right now is to play safe and stick to the 2.0
behavior in all cases beside the Unit where we still would use
ContextDependent (to have KT-71751 still fixed).

Thus, while for non-Unit cases we behave as before (in 2.0.20), for Unit
it's hardly possible that providing such an expected type might affect
callable references resolution in "natural" cases
(though I've found KT-73031, which is still not covered).

elvisOnJavaList.fir.kt and kt30796.kt have been reverted to the state
 they were before 20899ecaeba36cb327b543ce10da6c4fdc2556ba

^KT-72996 Fixed