[IC] Make FileSnapshotMap relocatable
Previously, only the map key was saved in a relocatable way, not the value. By removing `file` from `FileSnapshot`, we do not change correctness, because file path is used as the map key and we do not compare files from different paths. Additionally, this reduces the size of produced caches a bit.
^KT-70556 In Progress