[AA Platform] Add resolution scope enlarger for AA.
This allows individual extension points to register a callback to
enlarge the resolution scope of a `KaModule`. Unlike the existing
`KotlinResolutionScopeProvider`, multiple extension points can be
registered, and any resulting scopes will be unioned together.

This is designed to be the K2/AA equivalent of the K1/IDE
`KotlinResolveScopeEnlarger` extension point, allowing similar
functionality in IDE and non-IDE contexts.

^KT-73406 fixed