[JVM] Add hack for loading compiled internal inline function bytecode
When compiling code fragments from the debugger, the visibility rules
are weakened, and we can inline internal functions from other modules.
Internal function names are being mangled. When evaluating expressions,
FIR of accessed declarations is being set-up from sources and does not
have enough info to calculate mangled names. So, we implement a hacky
way of matching, ignoring mangling suffix. This is basically the same as
how Eval4J approach this problem, see JDIEval#findMethodOrNull in the
IntelliJ codebase.

^KT-72930: Fixed