[FIR] Rewrite `findStaticallyKnownSubtype()`
In `isErasedAllowForTypeWithIrrelevantMixin`, at some point we come to
`isSubtypeOfForSingleClassifierType` where we collect supertypes of
`DerivedSuper` and take their arguments (`TypeVariable(S)` and `D`) and
intersect them. This results in `(TypeVariable(S) & Any) & D`, for which
inference later deduces:

1. `(TypeVariable(S) & Any) & D <: B`
  - in `simplifyConstraintForPossibleIntersectionSubType` we split
    "our type variables" form other types, thus considering
    `TypeVariable(S) & Any` and `D` separately. Because `D <: Any`,
    we relax the constraints for our type variables, allowing them
    to be nullable
2. `TypeVariable(S) & Any <: B?`
  - adding a constraint to a DNN means adding its nullable version to
    the original variable, so
3. `TypeVariable(S) <: B?`

All this leads to deriving `DerivedSuper<out B?>` as the statically
known type. Informally, we are allowed to consider it
`BaseSuper<out B?>` because `BaseSuper<D>` already allows us to conclude
`BaseSuper<out B>` from it if we need it.

In `rawCastToStarProjection_Ok.kt`, the problem is the way we calculate
the bare type: we call `computeRepresentativeTypeForBareType` which,
in turn, does `doUnify` instead of the approach we introduce in the
previous commit (`doUnify` produces `ValueType.Optional<out Any>`).
This is a false changed, and it will be fixed later.

The constraints filtering is backed by the following tests:

- `compiler/testData/diagnostics/tests/resolve/overloadConflicts/kt55722.kt`
    - a false-positive diagnostic appears as we start inferring `B<*>`
      instead of `B<in T>`
- `compiler/fir/analysis-tests/testData/resolve/bareTypesWithFlexibleArguments.kt`
    - we start inferring `java/util/LinkedHashSet<out ft<T & Any, T>>`
      instead of `java/util/LinkedHashSet<out T>`

Note that `isCheckAndVarianceWithSubRec.kt` is not supported by this
change and is left out of scope.

Also note that `isCheckAndVarianceWithSmartcast.kt` is not supported by
this commit. This test requires additional logic for intersections.

Also, `isCheckAndVarianceWithRecursiveBound.kt` is a good example
showing a flaw in the current approach: when checking the type variables
against the RHS arguments, instead of adding one constraint with them
and the RHS type their bounds are checked separately against the
argument projections. This is because "just adding" such a constraint
would be like asking the CS to find solutions for the variables, not
make sure they hold for all possible solutions. But with recursive
bounds this approach becomes naive: the original variables may still
appear inside them and thus be further constrained by subsequently
inferred the subtype constraints, instead of making sure they hold for
all possible values of the variable or its bounds.

^KT-7972 Fixed