FIR: Refine contracts for postponed atoms properties
NB: !isPostponedBecauseOfAmbiguity meant both
`state == State.NOT_RESOLVED_YET` and `state == State.RESOLVED`,
but I believe the main idea of the affected if's is that we _should_
consider input types as trivial for freshly introduced callable
references.

As for `state == State.RESOLVED`, it doesn't really matter what
we would return for that state, because the atom is completed, so
these properties shouldn't even be used.
I think we might've had an assertion here, but on the other
hand it seems too much.