Fix wrong IrErrorCallExpression gen for annotation in convertToIrConstructorCall
In the middle of `KaFirCompilerFacility.compile()`, when the IR gen
part accesses the `annotations` of `Fir2IrLazyClass`, it tries to
generate the IR annotation calls for annotations of `Fir2IrLazyClass`.
In particular, it uses
`CallAndReferenceGenerator.convertToIrConstructorCall()`. When we have

- file `A` in a module `moduleA`
- its dependency file `B` in a module `moduleB`
- the class `B` (in file `B`) has an annotation `Foo` in a binary module
  `lib`
- and 'A' is the main target of the CodeGen,

the `session` of `CallAndReferenceGenerator` is the session for
`moduleA`. Since `moduleA` has only dependency `moduleB`, finding a
class symbol `Foo` using the `session` does not work. We have to use the
session for `moduleB`, because `moduleB` has dependency on `lib`.
Actually, it is nature that class `B` has an annotation `Foo` (not `A`),
so we have to find `Foo` from `B` (with the session for `moduleB`).

However, the current implementation does not correctly use the session.
It generates IrErrorCallExpression if finding the symbol based on the
session does not work, which can cause serious issues. For example, if
a Compose annotation was correct, and it was dropped because of the
IrErrorCallExpression, the generated code will not correctly work for
Compose UI.

This commits updates `convertToIrConstructorCall()` to correctly use the
module session based on the call site of the annotation call instead of
generated IrErrorCallExpression.