[K/N] Add debugCompilationDir option akin to -fdebug-compilation-dir
When using -Xdebug-prefix-map in K/N, Apple's linker (ld-prime) seems
to discard the produced debug information because of the empty
DW_AT_comp_dir value.

To mitigate that, there should be a way to modify this value.

This commit adds -Xbinary=debugCompilationDir=$dir option, similar
to -fdebug-compilation-dir in Clang. The compiler propagates the
specified value to DW_AT_comp_dir.

^KT-73306 Fixed