[LL FIR] Introduce FirIgnoredContractDescription
When the first statement of a function is a call to a user-defined
`contract` function, we generate a `FirLegacyRawContractDescription`
even though the function has no contract, and then remove it later.
With this change, we explicitly mark that this happens instead of simply
setting the field to null, so that lazy loading won't re-add the removed
description.

^KT-73259 Fixed