[JVM] Always use CLASS lambda compilation scheme for code fragment
IDEA used to run code fragment compiler with CLASS compilation scheme,
since such lambdas have pretty toString implementation. However, it's
unsafe. Consider the case when the code fragment refers some local
declaration in the project code. To have all the required info, the
file that contains the local declaration will be compiled together with
code fragment, and the compilation will run with CLASS scheme. If the
project is actually being compiled with INDY (which is default),
compiled code fragment will be binary incompatible with the compiled
project. See KT-73049 as example.

However, we still can safely compile lambdas defined in the code
fragment itself with an arbitrary scheme, so we keep on using CLASS.