[KLIB ABI] Disable IrMangledNameAndSignatureDumpHandler in IR text tests
This handler is useless for Kotlin/JVM backend because Kotlin/JVM
does not use signatures neither for linkage nor for serialization.

This handler became obsolete for KLIB-based backends due to
`KlibAbiDumpHandler`, which dumps not only signatures of public entities
in a KLIB file but also the full ABI shape of the KLIB file.

^KT-73171