[Gradle] GMDT: Fix Included Build Project Dependencies
Also make sure that import doesn't trigger compile tasks.
Including CInterops.

To achieve that we need to separate GMDT tasks inputs depending
on the context. For cases when Metadata KLIBs have to be produced
we need to make sure that GMDT task depends
on "metadataLocationsOfProjectDependencies".
This FileCollection enforces Metadata Compilation of project dependencies

However, for IDE we don't need that. Because IDE should analyze sources
by itself. But in both cases we need to consume PSM and infer source set
visibility.
For that reason `transformProjectDependencies` was introduced.
Its responsibility to prevent from collecting Metadata files
by simply falling back to emtpy collection.
IDE will not use it.

CInterops are exposed as secondary variants.
And they are consumed both for IDE and regular Klib compilation.
So for CInterops we disable projectDependencies transformation
for all cases.

^KT-71074 Verification Pending