[FIR] Revert `inner` modifier from typealias in incremental test-data
It was advised by aleksei.cherepanov@jetbrains.com some time ago,
and this modifier is not needed because RHS doesn't have outer type parameters
(top-level type aliases with inner RHS are supported).

Also, it simplifies nested typealias bug-fixing
because `inner` typealias is not going to be supported in the nearest future.