CLI: hardcode DOS_EPOCH
This use of GregorianCalendar was semi-deterministic because it didn't specify the timezone or locale. The value can change based on the user's system settings.
Additionally, instantiating a Calendar is relatively expensive, because a bunch of locale-specific data may be parsed. This is insignificant for a single execution, but can add up in something like a distributed build system.
Hardcoding the exact value solves both of these problems.