[IR] Replace usages of allParameters with parameters when trivially safe
i.e. when the receiver is IrSimpleFunction.

allParameters is meant to account for a bit "imaginary" parameter of a
constructor. But it also seems to have been used to simply get all
declared parameters of IrSimpleFunction.
To reduce the confusion, we try to consistently use `parameters` property
in that case.