[IR] Align typeArguments implementation with (value) arguments
It makes sense to express and use them uniformly, in this case
as a public MutableList.
This step also helps to initialize type argument count via information
from symbol, like it is done for value arguments. This was omitted before
as not strictly necessary, but otherwise value and type arguments are
not much different here.

KT-70057 related