[IR] Do not extract local delegates from the inline function
Doing that can lead to inconsistent IR. Local delegated property
consists of two elements: property and accessors to it.
Inside the accessor we have a reference to the property.
`LocalClassesInInlineLambdasLowering` can only extract the accessor
out of inline lambda, leaving the property in place. Because of this,
we have not entirely correct reference.

#KT-72439
#KT-72775