K2: report FUNCTION_CALL_EXPECTED error in 2.2 for synthetic properties without Java origin
Previously, we reported SYNTHETIC_PROPERTY_WITHOUT_JAVA_ORIGIN warning
for this kind of questionable synthetic properties.
In KT-64358 is planned not to resolve to such properties at all.
However, this may lead to worse IDE experience, so here
we've forbidden this situation by reporting FUNCTION_CALL_EXPECTED.
But we still resolve to such kind of properties leaving the mode
when we don't do it experimental.

#KT-72305 Fixed
Related to KT-64358