K2 CompilerFacility: Set up moduleName for FIR class
The existing `PsiRawFirBuilder` skips setting the `moduleName` of
`FirRegularClass`. As a result, on `MethodSignatureMapper` in the jvm
codegen backend, we have `null` `moduleName` for `Fir2IrLazyClass`, and
we always use the fallback JvmProtoBufUtil.DEFAULT_MODULE_NAME that is
"main".

This causes the difference between the name of a method for an
invocation and the defined method name, which causes a runtime
exception, in particular, Android Studio LiveEdit feature has the
real-world example.

This commit sets up `moduleName` of `FirRegularClass` in
`PsiRawFirBuilder` to above the issue.

^KT-73699