[IR] Do not consider "unimplemented" FOs for computing modality, etc
1. "Unimplemented" FOs are generated in case there is an abstract
function or property in one of the super types of the final class.
Such FOs are used as a last resort if there are no other non-abstract
FOs in a final class: the most specific of "unimplemented" FOs
is chosen. Otherwise, if there are still non-abstract FOs, the choice
should be made without taking into account "unimplemented" FOs.

2. There are two `createAndBindFakeOverride()` functions in
`IrFakeOverrideBuilder`. One of them already had this kind of filtering.
The fact that the other one didn't have it is likely due to oversight.

^KT-73171