IR Scripting: move script lowering to the scripting plugin
The motivation is mainly related to the REPL implementation (#KT-59854),
which needs to reside in the plugin in order to transfer some data
from the lowering to the compilation of the next snippet.
But it looks cleaner for the regular scripting too.
Also implement a new way to pass result field calculation to the
REPL runner (via IrAttribute), and drop the old one.
The changes in testdata are due to the different phase in which
the script lowering now runs.