[PL] Process only enqueued declarations, don't traverse IR modules
In certain compilation modes (example: producing Native caches),
the Kotlin/Native compiler may not automatically deserialize
top level declarations from IR modules. Instead, it just creates
empty IR files and allows the declarations to be deserialized later
and on demand (i.e., when a declaration is referenced from somewhere
else).

This can have a negative effect on the partial linkage:
- When an IR module is visited, there are no declarations in it yet.
- When declarations are deserialized, the module is not revisited.

So, a declaration becomes unprocessed.

To overcome this, it was decided to change the approach:
- Track all deserialized top-level declarations. Store the queue of
  such declarations right inside the PL engine.
- Allow the PL engine to process the queue. Trigger processing in
  the IR linker and in Kotlin/Native inline function resolver (when
  a new inline function has been deserialized).

^KT-73511