Tests: implement runner that stops after *2Ir conversion
... to test diagnostics reported only by conversion step. Required to
fix "jvm IR diagnostic" tests on scripts after moving the lowering
to the plugin, so it is executed before the backend which leads to
verification errors instead of diagnostics reporting.
Also restore the FIR counterpart of these tests.

Note that FIR part should probably eventually migrate to the tiered
infra, but since K1 tests are still required for scripting, the
migration is postponed for now.