[Analysis API] introduce `KaResolutionActivityTracker`
This service provides a bridge between the Analysis API, Kotlin IntelliJ
plugin and Java IntelliJ plugin. In particular, it is indented to adjust
the behavior of the Java resolver with `PsiFileEx.BATCH_REFERENCE_PROCESSING`
flag to avoid an aggressive cache strategy to comply with the Kotlin
compiler phase contracts when the Kotlin resolver is on the stack.
This service should be dropped as soon as the new platform API for
batch reference processing suppression will become available in
the Kotlin repo

^KT-73649