[PL] Remove problematic declarations from IR file w/o visiting the file
Now, given that the PL engine does not visit IR modules and IR files
anymore, it is not possible to schedule the problematic top-level
declarations for removal in their container (which is either IR file
or IR package fragment in case of Lazy IR).

With this fix, it becomes possible.

^KT-73511