[IR][Native] Add FirNativeSerializationTestGenerated for future testing of new nodes serialization
A new testrunner is created for Native serialization tests, which would shortly allow, in scope of KT-73171, to match IR before serialization and after deserialization.
Note: This testrunner may be considered as a prefix for the full codegen pipeline, though the following is still missing:
- implementation of deserializer
- backend invocation
- resulting executable invocation
Since Native backend is not invoked, the majority of // IGNORE_BACKEND: NATIVE directives are not applicable. So, another ignore directive is used in this new test runner: IGNORE_SERIALIZATION_TEST

^KT-73553 Fixed

Merge-request: KT-MR-19166
Merged-by: Vladimir Sukharev <Vladimir.Sukharev@jetbrains.com>