[JVM] Use `internal` visibility for $default-stubs of internal function
At the moment, we extend its visibility to public in IR. It is not a big
deal, but we still mangle those functions, and it breaks the
debugger-related hacks when we're looking up for mangled function
bytecode (see SourceCompilerForInline#getMethodNode).

Yet `internal` still will be `public` at JVM level, so it's safe.

^KT-73538: Fixed