K2: Run lazy-resolution on FIR value parameter before generating IR parameter
When K2 compiler facility AA generates IR for annotation for a value
parameter IR, the access to the resolved type of the annotation FIR
causes an exception. The exception stems from the FIR Fir2IrLazy value
parameter. The FIR is not fully resolved, which result in an exception
when accessing to the type of the annotation.

^KT-73843