JVM: generate class implementations in -Xjvm-default=all-compatibility
Generate the same set of methods that is generated in the
`-Xjvm-default=disable` mode. The only difference is that in the body,
we call a super-method instead of a method from the super-interface's
DefaultImpls.

Note that behavior has also changed in the `-Xjvm-default=disable` mode:
now we also generate those implementations in classes, which are
inherited from classes compiled in `all`/`all-compatibility` modes (see
the test `nonDefaultInheritance/specialization.kt`).

The change in `ExpressionCodegen` is needed to fix the case of
generating super calls from an inline value class, for example in the
test `defaultWithDefaultParameter/all_compatibility.kt`. Mapping
dispatch receiver type via `asmType` led to VerifyError because the
inline class instance was unboxed.

The test `inheritPropertyJvmDefaultAllCompatibility.kt` became muted
because now we're trying to generate an override for a property getter
with an invalid name, exactly like in the already existing test
`inheritProperty.kt`, which is muted because of KT-68452.

 #KT-73954 Fixed