[decompiler] treat multifile class part classes as non-synthetic in special cases
Problem: an inconsistency between stub builders for such classes.
`KotlinClsStubBuilder` considered such declarations synthetic and
doesn't build a stub tree, but `KotlinClassFileDecompiler` does.
In this case someone may request a FirDeclaration for a stubbed
declaration from the decompiled file, and stub-based deserialized
should find it in the original non-decompiled class.

An alternative solution: allow building stubs in both cases and exclude
them from indices explicitly. This is a more correct solution, but,
unfortunately, it is too expensive to filter out such cases on earliest
steps of indexing due to the access to the file content to check
the metadata.

^KT-68484 Fixed