[FIR] Fix visibility of private top-level declaration from dangling file
...when checked from KaVisibilityChecker.

The problem is that useSiteFile is from a dangling module, while
declaration's containing file is from its context module.
That's why simple `==` check is not enough, and we have to unwrap
useSiteFile.
In the case of code fragments, context file can be found by context
element. In case of dangling files in completion we rely on
`unwrapCopy()`.

Q: Why does special handling for KaVisibilityChecker is needed? It
worked fine in normal resolve.
A: During resolve in IGNORE_SELF mode the useSiteFile will be replaced
by the respective file from the source module by
`patchDesignationPathForCopy()`. But this doesn't happen in
`KaVisibilityChecker`.

^KTIJ-29375 Fixed