[Analysis API] Implement partial body analysis
Partial body analysis can be performed on block bodies of functions,
constructors and initializers, including anonymous initializers in
scripts.

The function body is split into topmost statements. E.g., in the
following function declaration, there are three statements
(a function call, an 'if' statement, and another function call):

```
fun test(flag: Bool) {
    before()
    if (flag) {
        sendReport()
    }
    after()
}
```

If only several statements need to be analyzed for certain functionality
(such as, Find Usages may run on `fun before()`), 'FirElementProvider'
will request partial body analysis.

In a special case when no statements need to be analyzed but default
parameter values, 'FirElementProvider' requires a partial analysis with
the '0' target statement count.

The transformer part is not required to perform strictly partial body
analysis. In fact, now, after several attempts of partial body analyses,
a full analysis happens. The motivation that if semantic data from a
declaration is already requested quite a few times, it is likely that
the rest parts of the body will also be analyzed.

Each partial analysis result is stable and final – even if a subsequent
analysis of next statements fails, the clients may still use the already
available data. It works through copying the analyzed statements in
state keepers and through the deep CFG graph copying.

Note that for expression bodies partial analysis is not performed.
Potentially, it can be allowed for built-in apply/let-like functions,
or maybe even for certain user-provided functions. However, most likely
it will quite significantly complexify the analysis while the
performance benefits are unclear.

^KT-72357 Fixed