[PowerAssert] Compare via operator should show only the boolean result
When a comparison is made between two instances using operator
overloading, power-assert will show two values: 1. the result of the
`compareTo` call, and 2. the comparison of that result with `0`. Both of
these are incorrectly placed in the final diagram, and the first is not
very useful. Instead, only show the boolean result and make sure it is
correctly placed in the diagram.

^KT-73871 Fixed