FIR: Fix almost infinite recursively incorporated constraints processing
See insideOtherConstraintsWithFlexibleTypes.kt

In the case of flexible types, the pattern of an incorporated
constraints introduction is changed a bit: we add some constraints
with non-flexible variables, on the next the same one with
their flexible versions.

Then, we start incorporating them between themselves.

Though the root problem is that we remembered the origin variable
for "Inside Other Constraints" kind to avoid recursion.

But when the fresh constraints created via incorporation of that kind
were being used for simpler "Direct With The Same Variable",
the information got lost, thus allowing creating recursively originated
constraints.

The idea of the fix is just preserving the information.

I don't expect it to break anything because such recursively natured
constraints likely result in either infinite processing or in
NEW_INFERENCE_ERROR of some kind.

^KT-73434 Fixed