[Build] Use context parameters instead of receivers
Starting with Kotlin 2.2, a deprecation warning is emitted when the
context receivers language feature is enabled (see KT-73146). Because
Kotlin is built with allWarningsAsErrors enabled, a migration to context
parameters is required. Thankfully, this migration is very small and
mostly impacts build files, which must now enable context parameters
instead of context receivers.

^KT-71792