[K/Wasm] Set the correct file entry for lambdas from inline functions that were not inlined
The commit is fixing the setting of debug information for the following situation:

```kotlin
// FILE: a.kt
fun main() {
  val lambda = inlined()
  lambda()
}

// FILE: b.kt
inline fun inlined() = { 1 + 2 }
```

Before those changes, the Kotlin/Wasm backend emitted debug information with the `a.kt` file as a source file for the lambda returned from the `inlined` function in the example above. Current changes fix such a case. However, the problem is still reproducible in Kotlin/Native, and to fix it, it could be reused the same newly introduced `sourceFileEntry` attribute as in Wasm.