FIR: Rework Resolution & Inference for callable references
Previously, we performed
`addOtherSystem(candidate.system.currentStorage())`
around `initializeCallableReferenceAdaptation` call via CS
of a containing call, which led to `initialConstraints`
duplication in the case of PCLA when both the containing call and
the reference candidate use shared outer CS.

That might be fixed via using some modification of `addOtherSystem`,
that would avoid duplication.

But I've chosen the approach which fixes the issue _and_ simplifies
the code: now, instead of playing with transactions, applying them back
and forth, for callable references, their CSs inherit a CS
from the containing call.

And after that, we only need to call `replaceContentWith` to integrate
the successful candidate back.

Also, the hack with
`it is ConeTypeVariableType && it.typeConstructor !in outerCsBuilder.currentStorage().allTypeVariables` check
has been removed as seemingly unnecessary.

^KT-73771 Fixed