[JVM] Simplify `PerformByIrFilePhase` by using `SimpleNamedCompilerPhase`
During this refactoring we dropped two operations on `PhaserState`:

1. `state.changePhaserStateType`
This operation is required to get rid of `stickyPostconditions`, but
it is an obsolete field, and it is not used anywhere meaningfully, so
it can be safely dropped.

2. Setting up `alreadyDone` field.
It is used only in combinations with parallel execution. Such execution
is used only for JVM codegen, contains a single phase and creates a
new `PhaserState` for each execution, so also can be safely dropped.

#KT-62953