[Test] Delete `AbstractKotlinCompilerTest.configuration` method
There were two different methods for test configuration:
1. `abstract fun TestConfigurationBuilder.configuration()`
2. `open fun configure(builder: TestConfigurationBuilder)`

The intent was the following:
- first (extension) method should be used to declare the configuration
  in the test
- second method is a workaround for cases when the test runner inherits
  not from the `AbstractKotlinCompilerTest`, but from some other runner
  with `configuration` implemented (it's impossible to call the `super`
  implementation of extension method)

The problem is that the number of cases when some test inherits another
test was underestimated, so the number of tests with overridden the first
method was comparable with tests with the second method overridden. Such
setup led to confusion and showed itself as error-prone. So it was decided
to completely remove the first (extension) method and leave only the
second one