Update parameter names in test data after ASM update
This is mostly a revert of the following commits:

1) 63ca8ebddb08d0cccf658828dd19708d4be3341f
2) eb3227381ba2a0d5e5f220cd6787ad848cbac96c

These commits were made to update intellij dependency (KTI-1416).
However, it seems that during that update, the asm dependency was not
updated. This led to a situation where the Kotlin compiler used intellij
with a fix for IDEA-331588:
https://github.com/JetBrains/intellij-community/commit/638600f3dbce6908b3513ea86305f8df096ae09e
But with an older ASM version that did not yet have
`ClassReader.VISIT_LOCAL_VARIABLES` and the corresponding code to handle
it.

Normally this would result in a NoSuchFieldError at runtime. However the
referenced field here is public static final, so it was inlined at the
call site, and everything worked as before, except the fact that
`ClsFileImpl` now passed `SKIP_CODE` (along with a value of
`VISIT_LOCAL_VARIABLES` = 256, which had no effect), which resulted in
parameter names of Java methods not being loaded.

Now that ASM is updated to a new version, `VISIT_LOCAL_VARIABLES`
started working as expected, and parameter names are loaded again.

Note that in most tests parameter names have been updated, but in some
diagnostic tests AA versions of tests have been muted, because behavior
differs between the compiler and AA modes (AA always uses compiled PSI).

 #KT-74276 Fixed