[Analysis API] Support multi-file dangling modules
Multi-file dangling modules are needed for code fragment compilation,
where the context is in the common module.

In that case, the 'KaFirCompilerFacility' creates a new dangling file
module, consisting of a target file and all its dependencies from the
context module, and sets the JVM implementing module as a new context.

Before, each dependency file was compiled separately, which not only
was inefficient, it also did not support sophisticated mutual calls
between inline functions.

Even though the initial case came from 'KaFirCompilerFacility',
multi-file dangling file modules can be created in arbitrary places.
To do that:

 1. Create all in-memory files (e.g., with a 'KtPsiFactory');
 2. Manually initiate a [KaDanglingFileModule], passing the files.
 3. Set the newly created module as an [KtFile.explicitModule] to all
    of the created files.