[IR] Support new reference nodes in klib serialization
This formally required ABI bump. We can't do it in .20 a version.
That's why, new nodes should be never serialized for stable klibs
until 2.2. But it's fine to be able to deserialize them.

Serialization can happen in unstable new inlining mode, which is fine,
as it already requires 2.2-abi features.

^KT-72931