[FIR2IR] Consider `declarationSiteSymbol` while applying receivers instead of using `qualifiedAccess.toResolvedCallableSymbol()`
Previously, it worked correctly for all cases except type alias constructor (even for fake overrides).
The latter is synthetic, and such a callable reference is always replaced
by a constructor call created in `TypeAliasConstructorsSubstitutingScope`.
That constructor could have either source or fake-override origin.
If it has a fake-override origin, the `dispatchReceiverType` could be null
 (when base constructor has type parameters).

In `applyReceivers` we should consider original (unwrapped)
callable references to properly handle `dispatchReceiverType`,
and probably other things.
Also, it eliminates excess `toResolvedCallableSymbol` calls.

The fix is needed for the next commit that fixes KT-74107